reorganized types
